Note submittal, storage, and NoteIDs working
Had to refactor the way I was storing the information for the notes.

It was just storing all 4 variables in 4 different strings. Now instead Its convering the localStorage item into a JSON string. Then on window load its parsing the string to make it into an array again. And displaying properly.

NoteIDs just needed the proper operators assigned