Redoing File Scheme, Completed Note Propigation
This build doesnt have saves working. After doing an audit on how it worked and the reusability of the code it was nearly non-exist ant. Versions of the same local storage instances saving during multiple portions was just a recipe for disaster.

This build has the editor functionality, or writing and propagating the data to the correct places.

The next commit to this build will have the save functionality redone, then it will go into main.